Handle parsing ZMQ messages in worker queues
