+++ coinpusher/lib/zmq/zmq_raw_tx.ex
@@ -1,0 +1,12 @@
+defmodule CoinPusher.ZMQRawTx do
+  alias CoinPusher.{RawTransaction, NotificationsController}
+
+  def parse(data) do
+		case RawTransaction.parse(data) do
+      {:ok, tx, <<>>} ->
+        NotificationsController.notify_transaction(tx)
+      {:error, reason} ->
+        IO.inspect reason
+    end
+  end
+end
